Simplify public interface for numeric value enums
The variants for Resolution and FrameRate are named after numeric values
(ex: Sixteen is a variant name for both of them) but before this it
wasn't possible to say `FrameRate::from(16)` as the discriminat values
were the internal values used in the camera registers. Now private
methods are used for conversion to/from the camera register values, and
the public traits can be used by library users.