Require both WriteRead and Write for the I2C implementation
Turns out the camera gets angry and throws an error the next time you
try accessing it if you do a 0-length read operation after a write.

If/when embedded-hal releases 1.0, I can replace this with a
Transactional implementation. It'll be more work, but being able to do
write+read or write depending on what I want will be easier for the
generics than this mess.