Split CalibrationData from MelexisCamera, use Iterators for per-pixel data
Large diff here, but it can be distilled down to splitting
CalibrationData from MelexisCamera, and changing the per-pixel methods
of CalibrationData to return Iterators of references. Splitting out the
caluculation functions into a separate module then made the most sense
to resolve lifetime issues.

Splitting CalibrationData out and making it part of the public interface
allows users to implement it themselves for cases where they may be
storing the calibration data in a different manner (ex: precomputing the
parameters and storing it in a microcontroller's EEPROM, or compressing
the computer data in some way). Using iterators for the per-pixel
methods complements this as now instead of requiring a concrete slice,
just a method of retrieving individual pixels in a stream is required.
This is already being used by the included Mlx90640Calibration to only
store the four K_V calibration constants that are repeated across all
pixels.

Having the iterators yield *references* to pixel data also requires
explicit lifetimes, which is complicated by Generic Associated Types not
being available in Rust (at least yet). The existing Camera
implementations would call each other, which lead to rustc complaining
about multiple mutable borrows. To fix this issue the calculation
functions were split into a separate module, and I'm starting to expose
them as part of the public API for the same reasons CalibrationData is
being exposed.