Add device-level mocking and consolidate test data
There is now a crate-level module, `test`, which contains the example
EEPROM dump (and is also where the example MLX909641 dump will go). It
also contains a mock I2C bus implementation (using the embedded_hal I2C
traits) that mocks out the entire device. The mock device is flexible
enough to also be used with the MLX09641 when support for that is
implemented as well. The embedded_hal_mock crate was removed from the
dependencies as well.