Define API for calculations module
Almost everything is public. The non-generic functions being exported
are also marked with `inline`. Those functions are quite small, and most
are also called in very hot loops, so inlining them across crate
boundaries should be quite helpful for performance. They are *not*
marked `inline(always)` though, to give consumers of the crate the
option to not inline them.