Implement Hamming code for MLX90641
This is a non-tradidtional implementation, as I was learning about
Hamming codes as I wrote this. Melexis moving all of the parity bits to
the end didn't help, as it didn't *look* like a normal Hamming code
until I rewrote the matrix with the parity bits in the normal locations.

As a bonus, the tests for these functions are run against the example
darta from the datasheet, and they found a spot where I had
mis-transcribed a value from the datasheet. It also found a value from
the datasheet that has an incorrect checksum!