Add full EEPROM test from Melexis for MLX90640
Melexis has a full example of an EEPROM dump and the final values to be
extracted from it in a spreadsheet in their C/C++ driver. This data was
very helpful as it exposed a couple of bugs I missed (as evidenced by
the previous few commits).

The spreadsheet also has examples of what is read from RAM, and what the
calculated temperatures should be. Those tests will be coming later.