Rewrite StatusRegister and ControlRegister to to tuple structs
Leaving the I2C register with separate fields, as it's not used as much
as the other two. Hopefully this can (ever so slightly) improve memory
usage, and it might improve performance for some of the tighter loops
(as instead of deciding the entire register at once, only the parts
needed are decided, piece by piece.